server/veronica/requirements.txt
requests>=2.31.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
boto3>=1.28.0
